First pass at implementing FastBias via the DC rack server.
This changes the Channel interface hierarchy a bit; Channel no longer has  a DacBoard or FpgaModel.
Those are pushed to FpgaChannel, and AdcChannel, SramChannel, StartDelayChannel, and
TimingChannel all extend that.
Meanwhile FiberChannel gets a DcRackFiberId (e.g. A, B, C, or D).

FastBiasChannel gets split into two subclasses, FastBiasFpgaChannel and FastBiasSerialChannel.
FastBiasFpgaChannel should turn out to be identical to the old FastBiasChannel.
FastBiasSerialChannel is dead simple for now: it takes a DcRackFiberId and a card
(i.e. 0, 1, 2, etc).
It has a getSetupPacket method, which is called during buildSequence.

Also added: FastBiasChannelBuilder.

User-facing changes:
There is a FastBiasSerial channel, which has the parameters [cardNumber, fiberChannel], e.g.:
('flux', ('FastBiasSerial', ['1', 'A']))

There is the "Serial Bias" setting, which takes the channel name and the voltage.